refactor(preview-actions): extract header composition and compaction helpers
